+++ blocklets/snap-kit/api/src/routes/index.ts
@@ -1,15 +1,12 @@
 import { crawlCode, crawlUrl, getLatestSnapshot, getSnapshot } from '@arcblock/crawler';
 import { Joi } from '@arcblock/validator';
+import { auth, session } from '@blocklet/sdk/lib/middlewares';
 import { Router } from 'express';
 import qs from 'querystring';
 
 import { logger } from '../libs/logger';
 
-const { session, auth } = require('@blocklet/sdk/lib/middlewares');
-
 const router = Router();
-
-router.use(session({ accessKey: true }));
 
 /**
  * Crawl page html
@@ -17,13 +14,13 @@ router.use(session({ accessKey: true }));
 const crawlSchema = Joi.object({
   url: Joi.string().uri().required(),
   headers: Joi.object().pattern(Joi.string(), Joi.string()).max(30),
-  timeout: Joi.number().integer().min(10).max(120).default(120),
+  timeout: Joi.number().integer().min(10).max(120).default(60),
   waitTime: Joi.number().integer().min(0).max(120).default(0),
   cookies: Joi.array().items(Joi.object({ name: Joi.string().required(), value: Joi.string().required() })),
   localStorage: Joi.array().items(Joi.object({ key: Joi.string().required(), value: Joi.string().required() })),
   sync: Joi.boolean().default(false),
 });
-router.post('/crawl', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.post('/crawl', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await crawlSchema.validateAsync(req.body);
 
   res.setTimeout(params.timeout * 1000, () => {
@@ -71,7 +68,7 @@ const crawlGetSchema = Joi.object({
   url: Joi.string().uri(),
 }).or('jobId', 'url');
 
-router.get('/crawl', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.get('/crawl', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await crawlGetSchema.validateAsync(req.query);
   const snapshot = params.jobId ? await getSnapshot(params.jobId) : await getLatestSnapshot(params.url);
 
@@ -94,7 +91,7 @@ const snapSchema = Joi.object({
   height: Joi.number().integer().min(500).default(900),
   quality: Joi.number().integer().min(1).max(100).default(80),
   format: Joi.string().valid('png', 'jpeg', 'webp').default('webp'),
-  timeout: Joi.number().integer().min(0).max(120).default(120),
+  timeout: Joi.number().integer().min(0).max(120).default(60),
   waitTime: Joi.number().integer().min(0).max(120).default(0),
   fullPage: Joi.boolean().default(false),
   headers: Joi.object().pattern(Joi.string(), Joi.string()).max(30),
@@ -102,7 +99,7 @@ const snapSchema = Joi.object({
   localStorage: Joi.array().items(Joi.object({ key: Joi.string().required(), value: Joi.string().required() })),
   sync: Joi.boolean().default(false),
 });
-router.post('/snap', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.post('/snap', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await snapSchema.validateAsync(req.body);
 
   res.setTimeout(params.timeout * 1000, () => {
@@ -148,7 +145,7 @@ router.post('/snap', auth({ methods: ['accessKey'] }), async (req, res) => {
 const snapGetSchema = Joi.object({
   jobId: Joi.string().required(),
 });
-router.get('/snap', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.get('/snap', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await snapGetSchema.validateAsync(req.query);
   const snapshot = await getSnapshot(params.jobId);
 
@@ -187,10 +184,10 @@ const carbonSchema = Joi.object({
   code: Joi.string().required(),
   format: Joi.string().valid('png', 'jpeg', 'webp').default('png'),
   sync: Joi.boolean().default(false),
-  timeout: Joi.number().integer().min(0).max(120).default(120),
+  timeout: Joi.number().integer().min(0).max(120).default(60),
 });
 
-router.post('/carbon', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.post('/carbon', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await carbonSchema.validateAsync(req.body);
   const { sync, timeout, ...carbonParams } = params;
 
@@ -235,7 +232,7 @@ router.post('/carbon', auth({ methods: ['accessKey'] }), async (req, res) => {
 const carbonGetSchema = Joi.object({
   jobId: Joi.string().required(),
 });
-router.get('/carbon', auth({ methods: ['accessKey'] }), async (req, res) => {
+router.get('/carbon', session({ accessKey: true }), auth({ methods: ['accessKey'] }), async (req, res) => {
   const params = await carbonGetSchema.validateAsync(req.query);
   const snapshot = await getSnapshot(params.jobId);