+++ packages/crawler/src/index.ts
@@ -9,7 +9,9 @@ import { migrate } from './store/migrate';
 
 export * from './crawler';
 export * from './services/snapshot';
+export * from './store/job';
 export * as utils from './utils';
+export * from './metrics';
 
 export async function initCrawler(
   params: Pick<Config, 'puppeteerPath' | 'siteCron' | 'cookies' | 'localStorage' | 'concurrency'>,